refactor: Split schedule page into components and fix type issues
